refactor: more typescript migrations
