+++ package.json
@@ -99,6 +99,7 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/react-hooks": "5.1.3",
     "@testing-library/user-event": "12.8.3",
+    "@types/color": "^4.2.0",
     "@wojtekmaj/enzyme-adapter-react-17": "0.8.0",
     "axios-mock-adapter": "1.22.0",
     "css-loader": "5.2.7",
