+++ src/components/settings/SettingsAppearanceTab/types.d.ts
@@ -1,0 +1,8 @@
+export type ColorAccessibilityChecker = (color: string, textColor: Color) => boolean;
+
+export type Theme = {
+  title: string,
+  button: string,
+  banner: string,
+  accent: string,
+};
